+++ python/tree/misc/docutree.py
@@ -14,6 +14,8 @@ from docutils.parsers import rst
 from docutils.parsers.rst import directives
 from docutils import statemachine
 import traceback
+import importlib
+import re
 
 # This file is a basic example of how to write a Sphinx plugin to document
 # custom data.
@@ -63,6 +65,56 @@ def _format_command(name, envvars, base=None):
     yield ''
 
 
+def _format_changelog(changes):
+    ''' Format a changlog for a Tree '''
+
+    for line in changes:
+        if 'Enviroment' in line or 'Changes' in line:
+            yield f'**{line}**'
+        elif re.search('[a-z]', line):
+            yield '* ' + _indent(line)
+        yield ''
+
+
+def load_module(module_path, error=None, products=None):
+    """Load the module."""
+
+    # Exception to raise
+    error = error if error else RuntimeError
+
+    # __import__ will fail on unicode,
+    # so we ensure module path is a string here.
+    module_path = str(module_path)
+    try:
+        module_name, attr_name = module_path.split(':', 1)
+    except ValueError:  # noqa
+        raise error('"{0}" is not of format "module:object"'.format(module_path))
+
+    # import the module
+    try:
+        mod = importlib.import_module(module_name)
+    except (Exception, SystemExit) as exc:
+        err_msg = 'Failed to import module "{0}". '.format(module_name)
+        if isinstance(exc, SystemExit):
+            err_msg += 'The module appeared to call sys.exit()'
+        else:
+            err_msg += 'The following exception was raised:\n{0}'.format(
+                traceback.format_exc())
+
+        raise error(err_msg)
+
+    # check what kind of module
+    if products:
+        module = mod.dm.products
+    else:
+        module = mod
+
+    if not hasattr(module, attr_name):
+        raise error('Module "{0}" has no attribute "{1}"'.format(module_name, attr_name))
+
+    return getattr(module, attr_name)
+
+
 class TreeDirective(rst.Directive):
     ''' The directive which instructs Sphinx how to format the Tree config documentation '''
 
@@ -73,34 +125,6 @@ class TreeDirective(rst.Directive):
         'title': directives.unchanged,
         'remove-sasbase': directives.flag,
     }
-
-    def _load_module(self, module_path):
-        """Load the module."""
-
-        # __import__ will fail on unicode,
-        # so we ensure module path is a string here.
-        module_path = str(module_path)
-
-        try:
-            module_name, attr_name = module_path.split(':', 1)
-        except ValueError:  # noqa
-            raise self.error('"{0}" is not of format "module:parser"'.format(module_path))
-
-        try:
-            mod = __import__(module_name, globals(), locals(), [attr_name])
-        except (Exception, SystemExit) as exc:  # noqa
-            err_msg = 'Failed to import "{0}" from "{1}". '.format(attr_name, module_name)
-            if isinstance(exc, SystemExit):
-                err_msg += 'The module appeared to call sys.exit()'
-            else:
-                err_msg += 'The following exception was raised:\n{0}'.format(traceback.format_exc())
-
-            raise self.error(err_msg)
-
-        if not hasattr(mod, attr_name):
-            raise self.error('Module "{0}" has no attribute "{1}"'.format(module_name, attr_name))
-
-        return getattr(mod, attr_name)
 
     def _generate_section(self, name, config, cfg_section='default', remove_sasbase=False):
         """Generate the relevant Sphinx nodes.
@@ -147,7 +171,7 @@ class TreeDirective(rst.Directive):
     def run(self):
         self.env = self.state.document.settings.env
 
-        command = self._load_module(self.arguments[0])
+        command = load_module(self.arguments[0])
 
         if 'prog' in self.options:
             prog_name = self.options.get('prog')
@@ -170,8 +194,61 @@ class TreeDirective(rst.Directive):
         return sections
 
 
+class TreeChangeDirective(rst.Directive):
+    has_content = False
+    required_arguments = 1
+    option_spec = {
+        'prog': directives.unchanged_required,
+        'title': directives.unchanged,
+        'drs': directives.unchanged_required,
+        'remove-sasbase': directives.flag,
+    }
+        
+    def run(self):
+        self.env = self.state.document.settings.env
+
+        command = load_module(self.arguments[0])
+
+        if 'prog' in self.options:
+            prog_name = self.options.get('prog')
+        else:
+            raise self.error(':prog: must be specified')
+
+        # get options
+        remove_sasbase = 'remove-sasbase' in self.options
+
+        # get drs
+        drs = self.options.get('drs')
+        new, old = drs.replace(' ', '').split(',')
+
+        # compute and format the changelog lines
+        lines = command(new, old, remove_sas=remove_sasbase)
+        title = lines[0]
+        lines = lines[1:]
+        lines = _format_changelog(lines)
+
+        # the source name
+        source_name = title
+
+        # Title
+        section = nodes.section(
+            '',
+            nodes.title(text=source_name),
+            ids=[nodes.make_id(source_name)],
+            names=[nodes.fully_normalize_name(source_name)])
+
+        result = statemachine.ViewList()
+        for line in lines:
+            result.append(line, source_name)
+
+        self.state.nested_parse(result, 0, section)
+
+        return [section]
+
+
 def setup(app):
     app.add_directive('datamodel', TreeDirective)
+    app.add_directive('changelog', TreeChangeDirective)